Cleaning up. Ensuring no errors and compatibility in layouts.
