+++ app/src/main/res/layout/widget_alarms_activity_list.xml
@@ -3,7 +3,7 @@
     android:id="@+id/alarm_list_top_layout"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:background="@color/primary_light"
+    android:background="#B3FF3D00"
     android:gravity="center"
     android:orientation="vertical">
 
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="40dp"
         android:background="#0090FF"
-        android:elevation="1dp"
+        android:baselineAligned="false"
         android:orientation="horizontal">
 
 
@@ -35,9 +35,9 @@
                 android:drawablePadding="4dp"
                 android:ellipsize="end"
                 android:maxLines="1"
-                android:text="Set alarm after:"
+                android:text="@string/set_alarm_after"
                 android:textColor="#ffffff"
-                android:textSize="14dp" />
+                android:textSize="14sp" />
 
         </LinearLayout>
 
@@ -50,6 +50,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:clickable="true"
+        android:focusable="true"
         android:orientation="vertical"
         android:paddingLeft="10dp"
         android:paddingRight="10dp">
@@ -60,7 +61,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="onClick"
-            android:text="  5 minutes"
+            android:text="@string/_5_minutes"
             android:textSize="30sp" />
 
         <TextView
@@ -68,7 +69,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="onClick"
-            android:text="10 minutes"
+            android:text="@string/_10_minutes"
             android:textSize="30sp" />
 
         <TextView
@@ -76,7 +77,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="onClick"
-            android:text="20 minutes"
+            android:text="@string/_20_minutes"
             android:textSize="30sp" />
 
         <TextView
@@ -84,7 +85,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="onClick"
-            android:text="  1 hour"
+            android:text="@string/_1_hour"
             android:textSize="30sp" />
 
         <TextView
@@ -92,7 +93,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="onClick"
-            android:text="  sleep goal"
+            android:text="@string/miband_prefs_fitness_goal"
             android:textSize="30sp" />
     </LinearLayout>
